Add lessonplan, classroom & assignment APIs
Add new APIs, repositories and domain models for LessonPlan, ClassroomTeaching and AssignmentMaster; introduce Dapper-based repository implementations and interfaces. Rename and standardize many domain properties (e.g. ClsId->ClassId, SecId->SectionId, txt->ClassName/SectionName/SubjectName, etc.) and update repository parameter mappings accordingly. Register new services in DI (ServiceCollectionExtensions) and add logging/validation to controllers. Update appsettings connection string and modify database SQL scripts (FetchProcedure, MergeProcedure, Tables) to align with the changes.

+++ API/SchoolApplication/SchoolApplication/Interface/IClassSectionAllocationRepository.cs
@@ -5,6 +5,6 @@ namespace SchoolApplication.Interface
     public interface IClassSectionAllocationRepository
     {
         Task<string> MergeAllocateClassSectionAsync(ClassSectionAllocation classSectionAllocation);
-        Task<(IEnumerable<ClassSectionAllocation> allocations, IEnumerable<SchoolInfo> SchoolDetails)> FetchClassSectionAllocationsAsync(short acadYearId);
+        Task<(IEnumerable<ClassSectionAllocation> allocations, IEnumerable<SchoolInfo> SchoolDetails)> FetchClassSectionAllocationsAsync(byte acadYearId);
     }
 }
