+++ API/SchoolApplication/SchoolApplication/Interface/IClassSectionSubjectMapRepository.cs
@@ -5,6 +5,6 @@ namespace SchoolApplication.Interface
     public interface IClassSectionSubjectMapRepository
     {
         Task<string> MergeClassSectionSubjectMapAsync(ClassSectionSubjectMap classSectionSubjectMap);
-        Task<(IEnumerable<ClassSectionSubjectMap> classSectionSubjectMaps, IEnumerable<SchoolInfo> SchoolDetails)> FetchClassSectionSubjectMapAsync(int classId, short acadYearId, int? sectionId);        
+        Task<(IEnumerable<ClassSectionSubjectMap> classSectionSubjectMaps, IEnumerable<SchoolInfo> SchoolDetails)> FetchClassSectionSubjectMapAsync(int classId, byte acadYearId, int? sectionId);        
     }
 }
